app/services/auth_service/pyproject.toml
[project]
name = "auth-service"
version = "1.0.0"
description = "jwt auth service, registrate, authenticate, logout, refresh"
authors = [
    {name = "pgod.sw",email = "pgod.sw@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standart] (>=0.34.2,<0.35.0)",
    "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "logger (>=1.4,<2.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=1.0.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "bcrypt (==4.0.1)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "cryptography (>=45.0.3,<46.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
